requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
flask
pypdf
python-dotenv
pandas
-e .

# pip install --upgrade langchain-pinecone==0.0.3
# pip install --upgrade pinecone-client